Remove computation boundary escapes from engine markets (#297)
## Summary
- remove computation-boundary escapes from the engine.markets package
- replace raw Double and raw conversion call-sites with opaque-type
algebra and add missing typed semantics where needed
- update affected economics and tests, keeping docs out of scope

+++ src/main/scala/com/boombustgroup/amorfati/config/CorpBondConfig.scala
@@ -40,5 +40,5 @@ case class CorpBondConfig(
     bankShare: Share = Share(0.30),
     ppkShare: Share = Share(0.15),
     recovery: Share = Share(0.30),
-    maturity: Double = 60.0,
+    maturity: Int = 60,
 )
